В рамках адаптации под 32/64 бита переписана почтовая система. Вообще по функционалу планы были больше, но в процессе в очередной раз появилась навязчивая идея таки прикрутить к маду какую-нить бд, а чем такие идеи заканчиваются мы знаем (ничем), поэтому лучше уж накатить что-то рабочее, а там уже разберемся. Изменения в почте: -пишется не бинарный файл, а в хмл -пишется не в открытом виде, чтобы было меньше проходящих мимо читателей, заодно вроде как экранирование пользовательского ввода -вместо ид юзаются уиды (и те, и другие заменены на int) -очистка писем с удаленными адресатами и системные старше года (это всякие сообщения о дикеях, возратах посылок и что там еще) -сохраняется раз в минуту по необходимости -там же идет вывод онлайн мессаг о новой почте тем, кому было что-то написано (чтобы нельзя было проспамить уведомлениями чара написанием писем в цикле, сейчас ему просто в течении минуты будет выдаваться одно сообщение, сколько бы писем ему там не написали за этот промежуток) -по show stats пишется текущее кол-во писем в ожидании -старая почта вроде сконверчена (правда после прохода через упомянутые выше фильтры из 1700+ мессаг остается что-то около 500) -небольшая защита от дурака: лимит на письма в ожидании, 10 для безмортовых чаров 1-5 уровня, 100 для всех остальных -письму ставится материал береста -для записи/чтения time_t воткнуты кодер/декодер в строковый формат исо8601, вроде это единственный нормально переносимый вариант, раз уж мы о переносимости, можно будет заюзать в других местах

+++ src/comm.h
@@ -58,7 +58,7 @@ int write_to_descriptor(socket_t desc, const char *txt, size_t total);
 void write_to_q(const char *txt, struct txt_q *queue, int aliased);
 void write_to_output(const char *txt, DESCRIPTOR_DATA * d);
 void string_add(DESCRIPTOR_DATA * d, char *str);
-void string_write(DESCRIPTOR_DATA * d, char **txt, size_t len, long mailto, void *data);
+void string_write(DESCRIPTOR_DATA * d, char **txt, size_t len, int mailto, void *data);
 
 int toggle_compression(DESCRIPTOR_DATA * d);
