+++ src/modify.h
@@ -8,7 +8,7 @@
 #include "structs.h"
 
 void string_add(DESCRIPTOR_DATA * d, char *str);
-void string_write(DESCRIPTOR_DATA * d, char **txt, size_t len, long mailto, void *data);
+void string_write(DESCRIPTOR_DATA * d, char **txt, size_t len, int mailto, void *data);
 void page_string(DESCRIPTOR_DATA * d, char *str, int keep_internal);
 void page_string(DESCRIPTOR_DATA * d, std::string buf);
 void print_con_prompt(DESCRIPTOR_DATA * d);
